+++ src/discljord/messaging.clj
@@ -887,3 +887,32 @@
   "Returns a promise containing the bot's OAuth2 application info."
   []
   [])
+
+;; -------------------------------------------------
+;; Stages
+
+(defendpoint create-stage-instance! nil
+  "Creates a new stage instance associated to a stage channel.
+
+  Returns a promise with the stage instance."
+  [channel-id topic]
+  [privacy-level])
+
+(defendpoint get-stage-instance! ::ds/channel-id
+  "Returns a promise containing the stage instance for the given channel."
+  []
+  [])
+
+(defendpoint modify-stage-instance! ::ds/channel-id
+  "Edits a stage instance for the given channel.
+
+  Returns a promise with the modified stage instance."
+  []
+  [topic privacy-level])
+
+(defendpoint delete-stage-instance! ::ds/channel-id
+  "Deletes a stage instance for the given channel.
+
+  Returns a promise of a boolean indicating whether the operation succeeded."
+  []
+  [])